+++ src/main/java/lotto/controller/RankCalculator.java
@@ -1,7 +1,11 @@
-package lotto.domain;
+package lotto.controller;
 
 import java.util.List;
 import java.util.stream.Collectors;
+
+import lotto.domain.Lotto;
+import lotto.domain.Rank;
+import lotto.domain.WinningNumber;
 
 public class RankCalculator {
 	public static List<Rank> calculateMultiple(List<Lotto> lottos, WinningNumber winningNumber) {
