+++ src/main/java/lotto/utils/NumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.utils;
 
 import java.util.List;
