+++ src/test/java/lotto/domain/RankCalculatorTest.java
@@ -11,6 +11,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import lotto.controller.RankCalculator;
+
 public class RankCalculatorTest {
 	@ParameterizedTest
 	@MethodSource("generateLottos")
